+++ tests/covalent_dispatcher_tests/_core/dispatcher_test.py
@@ -512,62 +512,7 @@ async def test_submit_initial_tasks(mocker):
 
 
 @pytest.mark.asyncio
-async def test_submit_task_group_single(mocker):
-    """Test submitting a singleton task groups"""
-    dispatch_id = "dispatch_1"
-    gid = 2
-    nodes = [2]
-
-    mock_get_abs_input = mocker.patch(
-        "covalent_dispatcher._core.dispatcher._get_abstract_task_inputs",
-        return_value={"args": [], "kwargs": {}},
-    )
-
-    mock_attrs = {
-        "name": "task",
-        "value": 5,
-        "executor": "local",
-        "executor_data": {},
-    }
-
-    mock_statuses = [
-        {"status": Result.NEW_OBJ},
-        {"status": Result.NEW_OBJ},
-        {"status": Result.NEW_OBJ},
-    ]
-
-    async def get_electron_attrs(dispatch_id, node_id, keys):
-        return {key: mock_attrs[key] for key in keys}
-
-    mocker.patch(
-        "covalent_dispatcher._core.dispatcher.datasvc.electron.get",
-        get_electron_attrs,
-    )
-
-    mocker.patch(
-        "covalent_dispatcher._core.dispatcher.datasvc.electron.get_bulk",
-        return_value=mock_statuses,
-    )
-
-    mocker.patch(
-        "covalent_dispatcher._core.dispatcher.datasvc.update_node_result",
-    )
-
-    # This will be removed in the next patch
-    mock_run_abs_task = mocker.patch(
-        "covalent_dispatcher._core.dispatcher.runner.run_abstract_task",
-    )
-
-    await _submit_task_group(dispatch_id, nodes, gid)
-    mock_run_abs_task.assert_called()
-    assert mock_get_abs_input.await_count == len(nodes)
-
-
-# Temporary only because the current runner does not support
-# nontrivial task groups.
-@pytest.mark.asyncio
-async def test_submit_task_group_multiple(mocker):
-    """Check that submitting multiple tasks errors out"""
+async def test_submit_task_group(mocker):
     dispatch_id = "dispatch_1"
     gid = 2
     nodes = [4, 3, 2]
@@ -607,13 +552,13 @@ async def test_submit_task_group_multiple(mocker):
         "covalent_dispatcher._core.dispatcher.datasvc.update_node_result",
     )
 
-    # This will be removed in the next patch
     mock_run_abs_task = mocker.patch(
-        "covalent_dispatcher._core.dispatcher.runner.run_abstract_task",
+        "covalent_dispatcher._core.dispatcher.runner_ng.run_abstract_task_group",
     )
 
-    with pytest.raises(RuntimeError):
-        await _submit_task_group(dispatch_id, nodes, gid)
+    await _submit_task_group(dispatch_id, nodes, gid)
+    mock_run_abs_task.assert_called()
+    assert mock_get_abs_input.await_count == len(nodes)
 
 
 @pytest.mark.asyncio
@@ -658,9 +603,8 @@ async def test_submit_task_group_skips_reusable(mocker):
         "covalent_dispatcher._core.dispatcher.datasvc.update_node_result",
     )
 
-    # Will be removed next patch
     mock_run_abs_task = mocker.patch(
-        "covalent_dispatcher._core.dispatcher.runner.run_abstract_task",
+        "covalent_dispatcher._core.dispatcher.runner_ng.run_abstract_task_group",
     )
 
     await _submit_task_group(dispatch_id, nodes, gid)
@@ -695,9 +639,8 @@ async def test_submit_parameter(mocker):
         "covalent_dispatcher._core.dispatcher.datasvc.update_node_result",
     )
 
-    # Will be removed next patch
     mock_run_abs_task = mocker.patch(
-        "covalent_dispatcher._core.dispatcher.runner.run_abstract_task",
+        "covalent_dispatcher._core.dispatcher.runner_ng.run_abstract_task_group",
     )
 
     await _submit_task_group(dispatch_id, [node_id], node_id)