+++ tests/covalent_tests/shared_files/utils_test.py
@@ -20,7 +20,7 @@
 import pytest
 
 from covalent._shared_files.config import get_config
-from covalent._shared_files.utils import filter_null_metadata, format_server_url
+from covalent._shared_files.utils import filter_null_metadata, format_server_url, get_named_params
 
 
 @pytest.mark.parametrize(
@@ -35,6 +35,18 @@ def test_filter_null_metadata(meta_dict, expected):
     """Test the filter null metadata function."""
     filtered = filter_null_metadata(meta_dict)
     assert filtered == expected
+
+
+def test_get_named_params():
+    """Tests the changes I made in covalent/covalent/_shared_files/utils.py for fixing ValueError in when using KEYWORD_ONLU parameter in electron func"""
+
+    def test_func(a, *, b):
+        return a + b
+
+    named_args, named_kwargs = get_named_params(test_func, [1], {"b": 2})
+
+    assert named_args == {"a": 1}
+    assert named_kwargs == {"b": 2}
 
 
 def test_format_server_url():